Added individual tests
