+++ warehouse/clickhouse/clickhouse_suite_test.go
@@ -1,13 +1,6 @@
 package clickhouse_test
 
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestPostgres(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "ClickHouse Suite")
-}
+//func TestClickhouse(t *testing.T) {
+//	RegisterFailHandler(Fail)
+//	RunSpecs(t, "ClickHouse Suite")
+//}
