+++ warehouse/deltalake/deltalake_test.go
@@ -6,11 +6,12 @@ import (
 	"fmt"
 	"github.com/cenkalti/backoff"
 	"github.com/gofrs/uuid"
+	"github.com/iancoleman/strcase"
+	"github.com/rudderlabs/rudder-server/config"
 	"github.com/rudderlabs/rudder-server/warehouse/client"
 	"github.com/rudderlabs/rudder-server/warehouse/deltalake"
 	"github.com/rudderlabs/rudder-server/warehouse/deltalake/databricks"
 	"github.com/rudderlabs/rudder-server/warehouse/testhelper"
-	"github.com/rudderlabs/rudder-server/warehouse/testhelper/util"
 	"log"
 	"os"
 	"strings"
@@ -68,20 +69,18 @@ func (*DeltalakeTest) EnhanceWorkspaceConfig(configMap map[string]string) {
 }
 
 func (*DeltalakeTest) SetUpDestination() {
-	DLTest.WriteKey = util.RandString(27)
+	DLTest.WriteKey = testhelper.RandString(27)
 	DLTest.Credentials = deltalakeCredentials()
 	DLTest.EventsMap = testhelper.EventsCountMap{
-		"identifies":    1,
-		"users":         1,
-		"tracks":        1,
-		"product_track": 1,
-		"pages":         1,
-		"screens":       1,
-		"aliases":       1,
-		"groups":        1,
-		"_groups":       1,
-		"gateway":       6,
-		"batchRT":       8,
+		"identifies": 1,
+		"users":      1,
+		"tracks":     1,
+		"pages":      1,
+		"screens":    1,
+		"aliases":    1,
+		"groups":     1,
+		"gateway":    6,
+		"batchRT":    8,
 	}
 	DLTest.TableTestQueryFreq = 5000 * time.Millisecond
 
@@ -100,7 +99,7 @@ func (*DeltalakeTest) SetUpDestination() {
 
 	backoffWithMaxRetry := backoff.WithMaxRetries(backoff.NewConstantBackOff(1*time.Second), uint64(5))
 	if err = backoff.Retry(operation, backoffWithMaxRetry); err != nil {
-		log.Panicf("could not connect to warehouse redshift with error: %s", err.Error())
+		log.Panicf("could not connect to warehouse deltalake with error: %s", err.Error())
 	}
 	return
 }
@@ -115,27 +114,63 @@ func TestDeltalake(t *testing.T) {
 
 	t.Parallel()
 
-	randomness := strings.ReplaceAll(uuid.Must(uuid.NewV4()).String(), "-", "")
+	verify := func() {
+		randomness := strings.ReplaceAll(uuid.Must(uuid.NewV4()).String(), "-", "")
 
-	whDestTest := &testhelper.WareHouseDestinationTest{
-		Client: &client.Client{
-			DBHandleT: DLTest.DB,
-			Type:      client.DBClient,
-		},
-		EventsCountMap:           DLTest.EventsMap,
-		WriteKey:                 DLTest.WriteKey,
-		UserId:                   fmt.Sprintf("userId_deltalake_%s", randomness),
-		Schema:                   "deltalake_wh_integration",
-		VerifyingTablesFrequency: DLTest.TableTestQueryFreq,
+		whDestTest := &testhelper.WareHouseDestinationTest{
+			Client: &client.Client{
+				DBHandleT: DLTest.DB,
+				Type:      client.DBClient,
+			},
+			EventsCountMap:           DLTest.EventsMap,
+			WriteKey:                 DLTest.WriteKey,
+			UserId:                   fmt.Sprintf("userId_deltalake_%s", randomness),
+			Event:                    fmt.Sprintf("Product Track %s", randomness),
+			Schema:                   "deltalake_wh_integration",
+			VerifyingTablesFrequency: DLTest.TableTestQueryFreq,
+		}
+		whDestTest.EventsCountMap[strcase.ToSnake(whDestTest.Event)] = 1
+
+		testhelper.SendEvents(t, whDestTest)
+		testhelper.VerifyingDestination(t, whDestTest)
+
+		randomness = strings.ReplaceAll(uuid.Must(uuid.NewV4()).String(), "-", "")
+		whDestTest.UserId = fmt.Sprintf("userId_deltalake_%s", randomness)
+		whDestTest.Event = fmt.Sprintf("Product Track %s", randomness)
+		whDestTest.EventsCountMap[strcase.ToSnake(whDestTest.Event)] = 1
+		testhelper.SendModifiedEvents(t, whDestTest)
+		testhelper.VerifyingDestination(t, whDestTest)
 	}
 
-	testhelper.SendEvents(t, whDestTest)
-	testhelper.VerifyingDestination(t, whDestTest)
+	t.Run("With Merge Mode With Partition", func(t *testing.T) {
+		config.SetString("Warehouse.deltalake.loadTableStrategy", "MERGE")
+		config.SetBool("Warehouse.deltalake.enablePartition", true)
+		deltalake.Init()
 
-	randomness = strings.ReplaceAll(uuid.Must(uuid.NewV4()).String(), "-", "")
-	whDestTest.UserId = fmt.Sprintf("userId_deltalake_%s", randomness)
-	testhelper.SendModifiedEvents(t, whDestTest)
-	testhelper.VerifyingDestination(t, whDestTest)
+		verify()
+	})
+	t.Run("With Append Mode With Partition", func(t *testing.T) {
+		config.SetString("Warehouse.deltalake.loadTableStrategy", "APPEND")
+		config.SetBool("Warehouse.deltalake.enablePartition", true)
+		deltalake.Init()
+
+		verify()
+	})
+
+	t.Run("With Merge Mode Without Partition", func(t *testing.T) {
+		config.SetString("Warehouse.deltalake.loadTableStrategy", "MERGE")
+		config.SetBool("Warehouse.deltalake.enablePartition", false)
+		deltalake.Init()
+
+		verify()
+	})
+	t.Run("With Append Mode Without Partition", func(t *testing.T) {
+		config.SetString("Warehouse.deltalake.loadTableStrategy", "APPEND")
+		config.SetBool("Warehouse.deltalake.enablePartition", false)
+		deltalake.Init()
+
+		verify()
+	})
 }
 
 func TestMain(m *testing.M) {
