+++ warehouse/init/init.go
@@ -1,0 +1,5 @@
+package init
+
+func InitDestinations() {
+	
+}
