+++ warehouse/redshift/redshift_suite_test.go
@@ -1,13 +1,6 @@
 package redshift_test
 
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestRedshift(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Redshift Suite")
-}
+//func TestRedshift(t *testing.T) {
+//	RegisterFailHandler(Fail)
+//	RunSpecs(t, "Redshift Suite")
+//}
