+++ warehouse/snowflake/snowflake_suite_test.go
@@ -1,13 +1,6 @@
 package snowflake_test
 
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestSnowflake(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Snowflake Suite")
-}
+//func TestSnowflake(t *testing.T) {
+//	RegisterFailHandler(Fail)
+//	RunSpecs(t, "Snowflake Suite")
+//}
