+++ warehouse/snowflake/snowflake_test.go
@@ -1,4 +1,4 @@
-package snowflake
+package snowflake_test
 
 import (
 	"database/sql"
@@ -7,9 +7,10 @@ import (
 	"fmt"
 	"github.com/cenkalti/backoff"
 	"github.com/gofrs/uuid"
+	"github.com/iancoleman/strcase"
 	"github.com/rudderlabs/rudder-server/warehouse/client"
+	"github.com/rudderlabs/rudder-server/warehouse/snowflake"
 	"github.com/rudderlabs/rudder-server/warehouse/testhelper"
-	"github.com/rudderlabs/rudder-server/warehouse/testhelper/util"
 	"log"
 	"os"
 	"strings"
@@ -69,19 +70,18 @@ func (*SnowflakeTest) EnhanceWorkspaceConfig(configMap map[string]string) {
 }
 
 func (*SnowflakeTest) SetUpDestination() {
-	SFTest.WriteKey = util.RandString(27)
+	SFTest.WriteKey = testhelper.RandString(27)
 	SFTest.Credentials = sfCredentials()
 	SFTest.EventsMap = testhelper.EventsCountMap{
-		"identifies":    1,
-		"users":         1,
-		"tracks":        1,
-		"product_track": 1,
-		"pages":         1,
-		"screens":       1,
-		"aliases":       1,
-		"groups":        1,
-		"gateway":       6,
-		"batchRT":       8,
+		"identifies": 1,
+		"users":      1,
+		"tracks":     1,
+		"pages":      1,
+		"screens":    1,
+		"aliases":    1,
+		"groups":     1,
+		"gateway":    6,
+		"batchRT":    8,
 	}
 	SFTest.TableTestQueryFreq = 100 * time.Millisecond
 
@@ -89,7 +89,7 @@ func (*SnowflakeTest) SetUpDestination() {
 
 	operation := func() error {
 		var err error
-		SFTest.DB, err = Connect(SnowflakeCredentialsT{
+		SFTest.DB, err = snowflake.Connect(snowflake.SnowflakeCredentialsT{
 			Account:  SFTest.Credentials.Account,
 			WHName:   SFTest.Credentials.Warehouse,
 			DBName:   SFTest.Credentials.Database,
@@ -126,15 +126,19 @@ func TestSnowflake(t *testing.T) {
 		EventsCountMap:           SFTest.EventsMap,
 		WriteKey:                 SFTest.WriteKey,
 		UserId:                   fmt.Sprintf("userId_snowflake_%s", randomness),
+		Event:                    fmt.Sprintf("Product Track %s", randomness),
 		Schema:                   "SNOWFLAKE_WH_INTEGRATION",
 		VerifyingTablesFrequency: SFTest.TableTestQueryFreq,
 	}
+	whDestTest.EventsCountMap[strcase.ToSnake(whDestTest.Event)] = 1
 
 	testhelper.SendEvents(t, whDestTest)
 	testhelper.VerifyingDestination(t, whDestTest)
 
 	randomness = strings.ReplaceAll(uuid.Must(uuid.NewV4()).String(), "-", "")
 	whDestTest.UserId = fmt.Sprintf("userId_snowflake_%s", randomness)
+	whDestTest.Event = fmt.Sprintf("Product Track %s", randomness)
+	whDestTest.EventsCountMap[strcase.ToSnake(whDestTest.Event)] = 1
 	testhelper.SendModifiedEvents(t, whDestTest)
 	testhelper.VerifyingDestination(t, whDestTest)
 }
