+++ warehouse/testhelper/transformer.go
@@ -3,7 +3,6 @@ package testhelper
 import (
 	"context"
 	"fmt"
-	"github.com/rudderlabs/rudder-server/warehouse/testhelper/util"
 	"strconv"
 	"time"
 )
@@ -14,7 +13,7 @@ func SetupTransformer() *TransformerResource {
 	url := fmt.Sprintf("%s/health", transformerEndPoint)
 
 	// Waiting until transformer is ready
-	util.WaitUntilReady(context.Background(), url, time.Minute, time.Second, "transformer")
+	WaitUntilReady(context.Background(), url, time.Minute, time.Second, "transformer")
 
 	return &TransformerResource{
 		Url:  transformerEndPoint,
