+++ warehouse/warehouse_test.go
@@ -1,0 +1,8 @@
+package warehouse
+
+import (
+	"testing"
+)
+
+func TestMain(m *testing.M) {
+}
